Order Change Cart

+++ online_book/myapp/views.py
@@ -1,14 +1,10 @@
 from django.shortcuts import render, HttpResponseRedirect
-from .models import book,Cart,Address,Order
+from .models import book,Cart,Address,Order, User
 from .forms import book_form, Signup_form
 from django.contrib import messages
 from django.db.models import Q
 from django.contrib.auth import authenticate, login, logout
 
-# Create your views here.
-# def Home(request):
-#     a = book.objects.all()
-#     return render(request, 'index.html', {'data':a})
 
 def find(request):
     if request.method == "POST":
@@ -56,17 +52,17 @@ def delete(request, id):
     
 def Cart_search(request):
     if request.user.is_authenticated:
+        cart1 = Cart.objects.filter(user=request.user)
+        count = cart1.count()
+        uname = request.user
+        b = Address.objects.all().values_list('id', flat=True)
+        print(count)
         if request.method == "POST":
             search = request.POST.get('search')
-            # all = book.objects.all()
             a = book.objects.filter(Q(Name__icontains = search) | Q(Author__icontains=search) | Q(Publisher__icontains=search) | Q(category__icontains=search))
         else:
             a = book.objects.all()
-            cart1 = Cart.objects.filter(user=request.user)
-            coun = cart1.count()
-            uname = request.user
-            b = Address.objects.all().values_list('id', flat=True)
-            return render(request, 'cart.html', {'data':a, 'cartcount':coun, 'uname':uname})
+        return render(request, 'cart.html', {'data':a, 'cartcount':count, 'uname':uname})
     else:
         return HttpResponseRedirect('/login/')
 
@@ -78,16 +74,29 @@ def add_to_cart(request):
             if not store:
                 items.quantity +=1
                 items.save()
+
             return HttpResponseRedirect('/cart/')
 
 def view_cart(request):
     if request.user.is_authenticated:
         user_cart = Cart.objects.filter(user=request.user).values_list('website_id', flat = True)
         cart1 = book.objects.filter(id__in = user_cart)
-
         qua= Cart.objects.filter(user=request.user)
         uname = request.user
-        return render(request, 'showcart.html', {'data':cart1,'quantity':qua, 'uname':uname})
+
+        amount = book.objects.filter(id__in = user_cart).values_list('Price', flat=True)
+        number_books = Cart.objects.filter(user=request.user).values_list('quantity', flat = True)
+        mul_amt_book = []
+        for i in range (0, len(amount)):
+            mul_amt_book.append(amount[i]*number_books[i])
+        add_amount = sum(mul_amt_book)
+
+        
+        a = Cart.objects.filter(user=request.user).values_list('quantity', flat=True)
+        total = 0
+        for i in range(0, len(a)):
+            total = total + a[i]
+        return render(request, 'showcart.html', {'data':cart1,'quantity':qua, 'uname':uname, 'total_items':total, 'add_amt': add_amount})
     else:
         return HttpResponseRedirect('/login/')
     
@@ -98,28 +107,31 @@ def Remove_cart(request, id):
         return HttpResponseRedirect('/viewcart/')
     
 def cart_increase(request):
-    if request.method == "GET":
-        cid = request.GET.get('cid')
-        items, store = Cart.objects.get_or_create(website_id = cid)
+    if request.user.is_authenticated:
+        if request.method == "GET":
+            cid = request.GET.get('cid')
+            items, store = Cart.objects.get_or_create(website_id = cid, user=request.user)
+            print(items)
 
-        if not store:
-            items.quantity += 1
-            items.save()
-        
-        return HttpResponseRedirect('/viewcart/')
+            if not store:
+                items.quantity += 1
+                items.save()
+            return HttpResponseRedirect('/viewcart/')
 
 def cart_decrease(request):
-    if request.method == "GET":
-        cid = request.GET.get('cid')
-        items,store = Cart.objects.get_or_create(website_id = cid)
+    if request.user.is_authenticated:
+        if request.method == "GET":
+            cid = request.GET.get('cid')
+            items,store = Cart.objects.get_or_create(website_id = cid,user=request.user)
+            
 
-        if not store:
-            items.quantity -= 1
-            items.save()
-            if items.quantity < 1:
-                items.delete()
+            if not store:
+                items.quantity -= 1
+                items.save()
+                if items.quantity < 1:
+                    items.delete()
 
-        return HttpResponseRedirect('/viewcart/')
+            return HttpResponseRedirect('/viewcart/')
 
 def Signup(request):
     if request.method == "POST":
@@ -149,17 +161,45 @@ def LogOut(request):
     return HttpResponseRedirect('/login/')
 
 def Order_place(request):
-    if request.user_is_authenticated:
+    if request.user.is_authenticated:
         if request.method == "POST":
             product_id = request.POST.get('pid')
-            cart_id = request.POST.get('cid')
-
-            Order.objects.create(user = request.user)
+            cart_id = Cart.objects.filter(user=request.user).values_list('id', flat=True)
+            print(cart_id)
+            Order.objects.create(user = request.user, Buy_direct_id=product_id, Buy_cart_id = cart_id)
+            return HttpResponseRedirect('/address/')
         
+def Order_place1(request):
+    if request.user.is_authenticated:
+        if request.method == "GET":
+            cart_id = Cart.objects.filter(user=request.user).values_list('website_id', flat=True)
+            cart_id1 = Cart.objects.filter(user=request.user).values_list('id', flat=True)
+
+            p_id = book.objects.filter(id__in=cart_id).values_list('id',flat=True)
+            for i in p_id:
+                print(i)
+                Order.objects.create(user = request.user, Buy_direct_id = i)
+            return HttpResponseRedirect('/address/')
+    
 
 def address_user(request):
     if request.user.is_authenticated:
+        order = Order.objects.filter(user=request.user).values_list('Buy_direct_id', flat=True)
+        order_c = Order.objects.filter(user=request.user).values_list('Buy_cart_id', flat=True)
+        print(order_c)
+
+        data = book.objects.filter(id__in=order) | book.objects.filter(id__in=order_c)
+
+        print(data)
+        name = User.objects.filter(id__in=order)
         if request.method == "POST":
+            street=request.POST.get('street')
             city = request.POST.get('city')
             pincode = request.POST.get('pincode')
-            city = request.POST.get('city')
+            state = request.POST.get('state')
+            product_id = request.POST.get('pid')
+            Address.objects.create(street=street,City=city, Pincode=pincode, State = state, user=request.user, product_id=product_id)
+        os = Address.objects.filter(user_id = request.user)    
+        return render(request, 'add.html', {'user_add':os, 'data':data})
+    else:
+        return HttpResponseRedirect('/login/')
